pyproject.toml
[project]
name = "genieapp"
dynamic = ["version"]
description = "Genie-powered data exploration app"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.37.0",
    "databricks-sdk>=0.74.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "ty>=0.0.12",
]

[tool.apx.metadata]
app-name = "genieapp"
app-slug = "genieapp"
app-entrypoint = "genieapp.backend.app:app"
api-prefix = "/api"
metadata-path = "src/genieapp/_metadata.py"

[tool.apx.ui]
root = "src/genieapp/ui"

[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
fallback-version = "0.0.0"

[tool.hatch.build.hooks.version]
path = "src/genieapp/_version.py"
template = '''
version = "{version}"
'''

[tool.hatch.build]
artifacts = ["src/genieapp/__dist__", "src/genieapp/_metadata.py"]
exclude = ["src/genieapp/ui"]
